+++ apps/workflows/checks.py
@@ -68,6 +68,7 @@ class CheckParser:
         return self._parse_string(check_desc)
 
     def _parse_string(self, check_desc):
+        original_check_desc = check_desc
         check_desc = check_desc.replace(" ", "_")
         # enable more human-readable negation
         if check_desc.startswith("is_not_"):
@@ -84,7 +85,10 @@ class CheckParser:
             self.desc2equals,
             self.desc2in,
         ]:
-            result = func(check_desc)
+            if func in [self.desc2equals, self.desc2not_equals]:
+                result = func(check_desc, original_check_desc)
+            else:
+                result = func(check_desc)
             if result is not None:
                 return result
 
@@ -155,19 +159,24 @@ class CheckParser:
 
                 return (doc, has_element)
 
-    def desc2equals(self, check_desc):
+    def get_original_check_description(
+        self, check_desc_after_underscore, original_check_desc
+    ):
+        total_chars = len(check_desc_after_underscore) * -1
+        original_check_desc = original_check_desc[total_chars:]
+        return original_check_desc
+
+    def desc2equals(self, check_desc, original_check_desc=None):
         """
         attribute to literal value equality check
 
-        currently only supports string attributes
-        which values do not contain any spaces
+        supports string attributes including those with spaces in values
         """
         if check_desc.count("_is_") == 1:
             attr, value = check_desc.split("_is_")
             attr = self.sanitize_attribute(attr)
 
             if hasattr(self.model, attr):
-                doc = f"check that {self.model.__name__} attribute {attr} has a value equal to {value}"
 
                 def choices_field(model, name):
                     """
@@ -184,6 +193,12 @@ class CheckParser:
                     or attr in self.TEXT_CHOICES_PROPERTIES
                 ):
                     value = value.upper()
+                elif original_check_desc:
+                    value = self.get_original_check_description(
+                        value, original_check_desc
+                    )
+
+                doc = f"check that {self.model.__name__} attribute {attr} has a value equal to {value}"
 
                 def compare_element(instance):
                     field = getattr(instance, attr)
@@ -191,7 +206,7 @@ class CheckParser:
 
                 return (doc, compare_element)
 
-    def desc2not_equals(self, check_desc):
+    def desc2not_equals(self, check_desc, original_check_desc=None):
         """
         negative attribute to literal value comparison check
 
@@ -201,7 +216,7 @@ class CheckParser:
         if check_desc.count("_not_is_") == 1:
             check_desc = check_desc.replace("_not_is_", "_is_")
 
-            result = self.desc2equals(check_desc)
+            result = self.desc2equals(check_desc, original_check_desc)
 
             if result is not None:
                 doc, func = result